refactor: rename registry_event_processor
Rename folder to realtime_message_processor
and module to message_processor, in line with
classes and other vars already renamed.
